minor code cleanup

+++ Blank_Ogre_Project/source/ogre_application.cpp
@@ -477,7 +477,7 @@ bool OgreApplication::frameRenderingQueued(const Ogre::FrameEvent& fe){
 	  overlay_manager.getOverlayElement( "MyTextArea_Score")->setCaption("Score: " + Ogre::StringConverter::toString(objectManager->getScore()));
 	  overlay_manager.getOverlayElement( "MyTextArea_Health")->setCaption("Health: " + Ogre::StringConverter::toString(objectManager->getPlayerHealth()));
 	  overlay_manager.getOverlayElement("MyTextArea_Rocket_CD")->setCaption("Rocket Cooldown: " + Ogre::StringConverter::toString(static_cast<int>(player->getRocketCD())));
-	  overlay_manager.getOverlayElement("MyTextArea_Laser_CD")->setCaption("Laser Cooldown: "  + Ogre::StringConverter::toString(static_cast<int>(player->getLaserCD())));
+	  overlay_manager.getOverlayElement("MyTextArea_Laser_CD")->setCaption("Laser Cooldown: "  + Ogre::StringConverter::toString(static_cast<double>(player->getLaserCD())));
 	  overlay_manager.getOverlayElement("MyTextArea_Scatter_CD")->setCaption("ScatterShot Cooldown: "  + Ogre::StringConverter::toString(static_cast<int>(player->getScatterCD())));
 	}else{
